mixin method extraction, inclusion in invoice line and synchro fixes
