+++ l10n_br_account/models/__init__.py
@@ -8,7 +8,6 @@ from . import fiscal_tax
 from . import fiscal_operation
 from . import fiscal_operation_line
 from . import fiscal_document
-from . import fiscal_document_line
 from . import account_invoice
 from . import account_invoice_line
 from . import account_move_line
