+++ l10n_br_fiscal/demo/fiscal_operation_demo.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo noupdate="0">
 
-    <record id="l10n_br_fiscal.fo_venda" model="l10n_br_fiscal.operation">
+    <record id="fo_venda" model="l10n_br_fiscal.operation">
         <field name="comment_ids" eval="[(6,0,[ref('l10n_br_fiscal.fiscal_comment_dummy')])]"/>
     </record>
 
-    <record id="l10n_br_fiscal.fo_venda_venda" model="l10n_br_fiscal.operation.line">
+    <record id="fo_venda_venda" model="l10n_br_fiscal.operation.line">
         <field name="comment_ids" eval="[(6,0,[ref('l10n_br_fiscal.fiscal_line_comment_dummy')])]"/>
     </record>
