+++ l10n_br_fiscal/models/document_cancel.py
@@ -3,7 +3,7 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from odoo import api, fields, models
-from odoo.addons.l10n_br_fiscal.constants.fiscal import (
+from ..constants.fiscal import (
     SITUACAO_EDOC_CANCELADA, SITUACAO_FISCAL_CANCELADO,
 )
